[temporary] Test staging site configuration
Temporarily removing a period from the end of a sentence in the alt text
of an image to verify that the branches and publishing of this repo are
set up correctly.

After this verification is done, we plan to revert this commit.

This change should only propagate to the `v0.1` version of the staging
docs

[#168857566]

Signed-off-by: Jaskanwal Pawar <jpawar@pivotal.io>
Co-authored-by: Jaskanwal Pawar <jpawar@pivotal.io>