TPG 1.0 restore changes for restore
TPG 1.0 has broken alot of the flow with velero, and given the team is winding down. We have opted for a manual flow based on the TPG docs to restore the DB's without using velero.

The backup procedure is still via velero. The UX isn't great, but this is where we are now.